requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg
pydantic
alembic
requests
python-multipart
email-validator
passlib[bcrypt]
python-jose[cryptography]
celery
redis
pytest
pydantic_settings
fastapi-mail
httpx
gevent
PyJWT
pydantic-settings
pytest-asyncio